* krcp.c, krlogin.c, krlogind.c, krsh.c, krshd.c, login.c,         setenv.c, v4rcp.c: Signed v.s unsigned int cleanup.
        * defines.h: rcmd_stream_{read,write} take size_t as length argument.

        * kcmd.c: Use GETSOCKNAME_ARG3_TYPE instead of assuming int. input
        and output handler take size_t as length argument instead of
        int. Other signed vs. unsigned fixes.

        * configure.in: Add KRB5_GETSOCKNAME_ARGS.

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@14050 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/appl/bsd/defines.h
@@ -34,8 +34,8 @@ extern int kcmd (int *sock, char **ahost, int /* u_short */ rport,
 		 enum kcmd_proto *protonum /* input and output */
 		 );
 
-extern int rcmd_stream_read (int fd, char *buf, int len, int secondary);
-extern int rcmd_stream_write (int fd, char *buf, int len, int secondary);
+extern int rcmd_stream_read (int fd, char *buf, size_t len, int secondary);
+extern int rcmd_stream_write (int fd, char *buf, size_t len, int secondary);
 extern int getport (int *);
 
 extern void rcmd_stream_init_krb5 (krb5_keyblock *in_keyblock,
